ml-service/requirements.txt
fastapi
uvicorn
numpy
opencv-python-headless
pillow
tensorflow-cpu
python-multipart
gdown